requirements.txt
Django>=2.0,<=3.0
psycopg2>=2.7,<2.9 
djangorestframework==3.11.0
djangorestframework_simplejwt==4.5.0
django-cors-headers==3.5.0
PyYAML==5.3.1
uritemplate==3.0.1
PyJWT==v1.7.1
gunicorn==20.1.0